backend/requirements.txt
Flask
Flask-Cors
PyJWT
gunicorn
google-generativeai
mysql-connector-python
python-dotenv
requests
countryinfo
